[monodroid] Allow forcing assemblies to be preloaded
On desktop, a default configuration DSO is used which, among others, sets the `uses_assembly_preload` flag to false.

Currently for normal app development, this value is ultimately driven by the `_AndroidEnablePreloadAssembliesDefault` MSBuild variable which is currently set to true meaning assemblies are automatically preloaded.

To allow to reproduce this same behavior on desktop (which is required for scenarios that scan the AppDomain for loaded assemblies like Forms renderer discovery or X.A ResourceIdManager value update logic), this commit adds a new parameter to `createNewContextWithData` which allows the caller to specify the behavior per-context and thus permit more flexibility in the future (e.g. if we don't see a need to preload we don't have to).